+++ next.config.js
@@ -2,7 +2,7 @@
 const nextConfig = {
   reactStrictMode: true,
   basePath: "/research/research-catalog",
-  pageExtensions: ["page.tsx", "page.ts", "page.jsx", "page.js"]
+  pageExtensions: ["page.tsx", "page.ts", "page.jsx", "page.js"],
 }
 
 module.exports = nextConfig
