+++ src/utils/utilsTests/bibUtils.test.tsx
@@ -1,4 +1,4 @@
-import { standardizeBibId } from "../../src/utils/bibUtils"
+import { standardizeBibId } from "../../utils/bibUtils"
 
 describe("bibUtils", () => {
   describe("standardizeBibId", () => {
